# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/projects": {
      "get": {
        "operationId": "ProjectsController_list",
        "summary": "Projects in the caller’s jurisdiction, with parcel counts by stage",
        "parameters": [
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "DRAFT",
                "SUBMITTED",
                "UNDER_SCRUTINY",
                "APPROVED",
                "ACTIVE",
                "ON_HOLD",
                "COMPLETED",
                "CLOSED"
              ],
              "type": "string"
            }
          },
          {
            "name": "stateCode",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "q",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "projects"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "post": {
        "operationId": "ProjectsController_create",
        "summary": "Create a project (draft)",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProjectDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "projects"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/projects/{id}": {
      "get": {
        "operationId": "ProjectsController_get",
        "summary": "Project detail: stage breakdown, notices, compensation totals",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "projects"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/proposals": {
      "get": {
        "operationId": "ProposalsController_list",
        "summary": "Acquisition proposals in the caller’s jurisdiction",
        "parameters": [
          {
            "name": "projectId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "proposals"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "post": {
        "operationId": "ProposalsController_create",
        "summary": "Create a draft proposal",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProposalDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "proposals"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/proposals/{id}": {
      "get": {
        "operationId": "ProposalsController_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "proposals"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/api/rr/entitlements": {
      "get": {
        "operationId": "RRController_entitlements",
        "summary": "Catalogue of R&R entitlements (Second Schedule)",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "rr"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/rr/families": {
      "get": {
        "operationId": "RRController_families",
        "parameters": [
          {
            "name": "projectId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "rr"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/rr/cases": {
      "get": {
        "operationId": "RRController_cases",
        "summary": "R&R cases with family and entitlement grants",
        "parameters": [
          {
            "name": "projectId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "IDENTIFIED",
                "ELIGIBILITY_VERIFIED",
                "PLAN_CREATED",
                "BENEFIT_ASSIGNED",
                "BENEFIT_DELIVERED",
                "COMPLETED"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "rr"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/rr/cases/{id}/grants": {
      "post": {
        "operationId": "RRController_assign",
        "summary": "Assign entitlements to a family (PLAN_CREATED → BENEFIT_ASSIGNED)",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignGrantsDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "rr"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "R-NLAM API",
    "description": "Real-Time National Land Acquisition & Management System. All routes need a Bearer token unless marked public. In AUTH_MODE=dev, get one from POST /api/auth/dev-login. Money is integer paise; dates are UTC.",
    "version": "2.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateProjectDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "example": "NH-753J-MH"
          },
          "name": {
            "type": "string",
            "example": "Wardha–Yavatmal 4-lane Highway"
          },
          "sector": {
            "type": "string",
            "example": "Roads"
          },
          "description": {
            "type": "string"
          },
          "stateCode": {
            "type": "string",
            "example": "MH"
          },
          "stateName": {
            "type": "string",
            "example": "Maharashtra"
          },
          "districtCodes": {
            "example": [
              "MH-WRD"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "districtNames": {
            "example": [
              "Wardha"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "piaName": {
            "type": "string",
            "example": "National Highways Authority of India"
          },
          "requiredAreaHa": {
            "type": "number",
            "example": 182.5
          },
          "estimatedCostRupees": {
            "type": "number",
            "example": 4200000000,
            "description": "Rupees (converted to paise server-side)"
          }
        },
        "required": [
          "code",
          "name",
          "sector",
          "stateCode",
          "stateName",
          "districtCodes",
          "districtNames",
          "piaName",
          "requiredAreaHa",
          "estimatedCostRupees"
        ]
      },
      "CreateProposalDto": {
        "type": "object",
        "properties": {
          "projectId": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "landRequiredHa": {
            "type": "number",
            "example": 42.5
          },
          "remarks": {
            "type": "string"
          }
        },
        "required": [
          "projectId",
          "title",
          "landRequiredHa"
        ]
      },
...
      "AssignGrantsDto": {
        "type": "object",
        "properties": {
          "entitlementCodes": {
            "example": [
              "HOUSE_RURAL",
              "SUBSISTENCE_GRANT"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "entitlementCodes"
        ]
      },
...
    }
  }
}